Refactor list eval tests
